+++ anyblok_pyramid/scripts.py
@@ -9,29 +9,13 @@
 from wsgiref.simple_server import make_server
 from anyblok.blok import BlokManager
 from anyblok.scripts import format_configuration
-from anyblok.registry import RegistryManager
 from anyblok.config import Configuration
 from .pyramid_config import Configurator
-from anyblok_pyramid.release import version
 import sys
 from anyblok import load_init_function_from_entry_points
 from .common import preload_databases
 from logging import getLogger
 logger = getLogger(__name__)
-
-
-Configuration.applications.update({
-    'pyramid': {
-        'prog': 'AnyBlok simple wsgi app, version %r' % version,
-        'description': "WSGI for test your AnyBlok / Pyramid app",
-        'configuration_groups': ['config', 'database'],
-    },
-    'gunicorn': {
-        'prog': 'AnyBlok gunicorn wsgi app, version %r' % version,
-        'description': "GUNICORN for test your AnyBlok / Pyramid app",
-        'configuration_groups': ['gunicorn', 'database'],
-    },
-})
 
 
 def anyblok_wsgi(application, configuration_groups, **kwargs):
@@ -41,14 +25,14 @@ def anyblok_wsgi(application, configuration_groups, **kwargs):
     :param \**kwargs: ArgumentParser named arguments
     """
     format_configuration(configuration_groups, 'preload', 'pyramid-debug',
-                         'wsgi', 'beaker')
+                         'wsgi')
     load_init_function_from_entry_points()
     Configuration.load(application,
                        configuration_groups=configuration_groups, **kwargs)
     BlokManager.load()
-    RegistryManager.add_needed_bloks('pyramid')
     config = Configurator()
     config.include_from_entry_point()
+    config.load_config_bloks()
 
     wsgi_host = Configuration.get('wsgi_host')
     wsgi_port = int(Configuration.get('wsgi_port'))
@@ -72,8 +56,7 @@ def gunicorn_anyblok_wsgi(application, configuration_groups, **kwargs):
         logger.error("No gunicorn installed")
         sys.exit(1)
 
-    format_configuration(configuration_groups, 'preload', 'pyramid-debug',
-                         'beaker')
+    format_configuration(configuration_groups, 'preload', 'pyramid-debug')
     from .gunicorn import WSGIApplication
     WSGIApplication(application,
                     configuration_groups=configuration_groups).run()